test(button): ionic theme tests (#29201)
Issue number: N/A

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Adding new e2e tests into Ionic Framework can be difficult for new
contributors when it comes to knowing where existing test cases exist
and how to additionally add a new theme or option to the test.

To avoid a lot of back and forth feedback in Github comments, I split up
the test on the original PR to help demonstrate how to match existing
test structures.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Removes the proposed test specific to the Ionic theme.
- Migrates the tests into the existing e2e files for button, adding
specific test cases that only apply to the ionic theme.
- Added screenshot references

In this work, I tried to correct some legacy folder naming:
- `button/test/round` → `button/test/shape`
- For `fill="clear"` I updated the screenshot names, but left the folder
structure. There is more work here to organize `fill="outline"` and it
is not required at this time.


## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!--
  If this introduces a breaking change:
1. Describe the impact and migration path for existing applications
below.
  2. Update the BREAKING.md file with the breaking change.
3. Add "BREAKING CHANGE: [...]" to the commit description when merging.
See
https://github.com/ionic-team/ionic-framework/blob/main/.github/CONTRIBUTING.md#footer
for more information.
-->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

I left the HTML template for the ionic theme test alone. I figure this
is helpful to the developer, until the primary PR is approved. Then it
could be removed if it is no longer useful.